[BE] Fix: diary-service 유저번호 X-User-Seq 사용

+++ BE/diary-service/src/main/java/com/c202/diary/diary/entity/Diary.java
@@ -15,7 +15,7 @@ public class Diary {
     private int diarySeq;
 
     @Column(nullable = false)
-    private int userSeq;
+    private Long userSeq;
 
     @Column(nullable = false, length = 50)
     private String title;
